Added counting simple resource types

+++ Assets/Tasks/EverydayTasks/EdTasksC.cs
@@ -1,6 +1,6 @@
 public class EdTasksC
 {
-    public float chanceToLose, chanceToLost, chanceToRecruit, ownTrustAdd;
+    public int chanceToLose, chanceToLost, chanceToRecruit, ownTrustAdd;
     public int moneyAdd, trustAdd, evilScoreAdd;
     public string description;
 }
